Use engelsystem v.7.0.0.
+ Changelog: https://github.com/johnjohndoe/engelsystem/blob/master/CHANGELOG.md#v700.
+ Refactor tests to invoke EngelsystemNetworkRepository#load directly.
