+++ buildSrc/src/main/kotlin/nerd/tuxmobil/fahrplan/congress/Dependencies.kt
@@ -49,7 +49,7 @@ object Libs {
         const val coreKtx = "1.7.0"
         const val coreTesting = "2.1.0"
         const val emailIntentBuilder = "2.0.0"
-        const val engelsystem = "6.0.0"
+        const val engelsystem = "7.0.0"
         const val espresso = "3.4.0"
         const val junit = "4.13.2"
         const val kotlinCoroutines = "1.6.0"
