+++ engelsystem/src/main/java/info/metadude/android/eventfahrplan/engelsystem/EngelsystemNetworkRepository.kt
@@ -1,26 +1,25 @@
 package info.metadude.android.eventfahrplan.engelsystem
 
-import info.metadude.android.eventfahrplan.engelsystem.loading.ShiftsLoading.awaitShiftsResult
 import info.metadude.android.eventfahrplan.engelsystem.models.ShiftsResult
 import info.metadude.android.eventfahrplan.engelsystem.utils.UriParser
-import info.metadude.kotlin.library.engelsystem.ApiModule
 import okhttp3.OkHttpClient
-import java.net.URISyntaxException
+import retrofit2.HttpException
 
 class EngelsystemNetworkRepository(
 
-        private val uriParser: UriParser = UriParser()
+    private val uriParser: UriParser = UriParser(),
+    private val serviceProvider: EngelsystemServiceProvider = EngelsystemServiceProvider.getNewInstance()
 
 ) {
 
     suspend fun load(okHttpClient: OkHttpClient, url: String) = try {
         val uri = uriParser.parseUri(url)
-        val service = ApiModule.provideEngelsystemService(uri.baseUrl, okHttpClient)
-        val call = service.getShifts(uri.pathPart, uri.apiKey)
-        call.awaitShiftsResult()
-    } catch (e: URISyntaxException) {
-        ShiftsResult.Exception(e)
-    } catch (e: IllegalArgumentException) {
+        val service = serviceProvider.getService(uri.baseUrl, okHttpClient)
+        val shifts = service.getShifts(uri.pathPart, uri.apiKey)
+        ShiftsResult.Success(shifts)
+    } catch (e: HttpException) {
+        ShiftsResult.Error(e.code(), e.message())
+    } catch (e: Exception) {
         ShiftsResult.Exception(e)
     }
