+++ engelsystem/src/main/java/info/metadude/android/eventfahrplan/engelsystem/models/ShiftsResult.kt
@@ -9,12 +9,6 @@ sealed class ShiftsResult {
 
     open class Exception(val throwable: Throwable) : ShiftsResult() {
 
-        object MissingResponseSuccessBody : Exception(
-                NullPointerException("Response success body is null."))
-
-        object MissingResponseHttpStatusMessage : Exception(
-                NullPointerException("Response HTTP status message is null or empty."))
-
         override fun toString() = "Exception(throwable=$throwable)"
 
     }
